add Controller
게시글을 위한 Controller 추가
